Split Transform and RawSQL decorators (#425)
Closes: #400
